6-size.c

+++ 0x00-hello_world/6-size.c
@@ -1,4 +1,5 @@
 #include <stdio.h>
+
 /**
  * main - Entry point
  *
@@ -6,12 +7,13 @@
  *
  * Return: Always  0 (Success)
 */
+
 int main(void)
 {
-        printf("size of a char: %u byte(s)\n", (unsigned long)sizeof(char));
-        printf("size of an int: %u byte(s)\n", (unsigned long)sizeof(int));
-        printf("size of a long int: %u byte(s)\n", (unsigned long)sizeof(long int));
-        printf("size of a long long int: %u byte(s)\n", (unsigned long)sizeof(long long int));
-        printf("size of a float: %u byte(s)\n", (unsigned long)sizeof(float));
+        printf("size of a char: %u byte(s)\n", sizeof(char));
+        printf("size of an int: %u byte(s)\n", sizeof(int));
+        printf("size of a long int: %u byte(s)\n", sizeof(long int));
+        printf("size of a long long int: %u byte(s)\n", sizeof(long long int));
+        printf("size of a float: %u byte(s)\n", sizeof(float));
         return (0);
 }
